Change table notation to that in the ICFP paper (modulo order and unique clauses).
Table clauses are now declarations that must appear at top level.
Referring to the table by name in a Links expression is equivalent to
writing the table clause in place (in the system prior to this
changeset application).  This is implemented by a compulsory
"optimiser" pass that substitutes table declarations wherever their
names appear in the program.

See the examples for examples.

I expect that this, like quite a few things, doesn't work properly in
the interactive loop.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@158 f10b8b52-1f0f-0410-b9dc-dace7d27bc78