Polymorphic recursion.
Previously: all functions had to be used monomorphically within their
own definitions (and within the recursive group).

Now: if you give a type signature, polymorphic recursion is allowed.
The type signature is needed because while checking is fairly
straightforward, inference is undecidable.

See tests/polymorphism.tests for some silly examples.

This currently seems to be pretty useless since we have no way to
define nested datatypes: our mu-bound variables are all nullary, which
ensures regularity of the type.  If we had something like mu-bound
functors (unary type constructors) then this might be more useful.  If
we solve this, we'll be able to write code equivalent to the following
(valid) Haskell, currently impossible in Links:

    data Nest a = Nil | Cons a (Nest (a,a))

    sum :: (a -> Int) -> Nest a -> Int
    sum Nil = 0
    sum (Cons a rest) = f a + sum (double f) rest
       where double f (x,y) = f x + f y

    sum id (Cons 1 (Cons (2,3) (Cons ((4,5),(6,7)) Nil)))

Ultimately this might help with writing type-indexed generic
functions.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@461 f10b8b52-1f0f-0410-b9dc-dace7d27bc78