Tidy pattern parsing a bit.
We used to parse a pattern as follows:

   1. Parse patterns as expressions: have the parser generate a Sugar.phrase
   2. Convert the Sugar.phrase to a Syntax.untyped_expression
   3. Convert the Syntax.untyped_expression to a Sugar.pattern
   4. Compile the Sugar.pattern into a Syntax.expression

The new scheme is marginally less Gordian:

   1. Parse patterns as patterns: have the parser generate a Sugar.pattern
   2. Simplify n-ary bits of the pattern to binary bits, generating a Sugar.simple_pattern
   3. Compile the Sugar.simple pattern into a Syntax.expression

Ideally we should skip step 2 as well, but that's for another day.

Parsing patterns as patterns, rather than as expressions, gives a more
precise grammar (no `_' or `as' in the expression grammar; no
arithmetic etc. in the pattern grammar), but I'm not really sure that
it's the best thing overall.  There are at least two advantages to
doing it the old way:

 1. potentially better error messages.  Instead of "Syntax error at
    position ..." we could give "The following cannot appear in a
    pattern: ...".  This is hypothetical, though: it should still be
    possible to change the parser to get reasonable diagnostics.

 2. Fewer traversal functions.  We now need to traverse both patterns
    and phrases in order to collect type vars, for example.  If only
    we could automatically generate fold functions for datatypes...

Another change: `Sugar' now has a (narrow) interface (.mli file).  To
avoid the duplication of type declarations in the interface and the
implementation all type declarations are in a separate file,
sugartypes.mli.  This might be a reasonable strategy in other areas of
the code where we want to export datatypes.

I've split the code in sugar.ml into modules to indicate the two major
areas of functionality: converting phrases to syntax trees (AST to IR,
if you like) and the pattern compiler.  This has the advantage of
decreasing the scope of most names, making the code easier to
understand.  I'm not really sure whether I like it or not.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@478 f10b8b52-1f0f-0410-b9dc-dace7d27bc78