Generate Functor instances (map for parameterized datatypes) automatically.
We can generate functor instances for a type constructor if it has kind

   * -> *

and if any occurrences of the type variable on the rhs of the
definition occur either "alone", in a tuple, or as the argument to
type constructors that are themselves instances of Functor.  For
example,

   type 'a foo = 
     [ `A of ('a * int) | `B of (('a * 'a option) list list * 'a foo)]

is ok, since all of the occurrences of "'a" satisfy those criteria
(option and list are instances of Functor).  In constrast,

   type 'a bar = 
     [ `C of ('a * int) | `D of (('a ref * 'a option) list list * 'a bar)]

is not ok since the second occurrence of "'a" is as the argument of
`ref', which is not an instance of Functor (see
deriving/lib/functor.ml).

This means that we can generate Syntax.redecorate (which is just
`map') automatically.  The real motivation is to be able to generate
rewrite instances, though.  That'll follow shortly, if all goes well.

Instanecs for mutually-recursive types, records and type constructors
with multiple parameters not yet handled.  The latter seem
particularly interesting.

Also an implementation of Haskell's Monad library (Control.Monad),
with instances for `option' and `list'.  (This is relevant because
Monad a => Functor a, so if we have a monad instance for a type we get
a Functor instance for free).



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@523 f10b8b52-1f0f-0410-b9dc-dace7d27bc78