Fixed insertion of multiple rows in the postgres driver
In SQL:

insert into table(field_names) values (vs1), (vs2), ..., (vsn)
==
insert into table(field_names) values
 select vs1
 union
 select vs2
 union
 ...
 union
 vsn

Postgres supports the latter syntax, but not the former.

Unfortunately this doesn't work for Sqlite which doesn't support
insertion of multiple rows or insertion of the values returned by a
select. We could implement insertion of multiple rows in Sqlite by
wrapping a transaction around the insert... but this may not be
desirable as the way Sqlite implements transactions is to lock the
entire database when it encounters a 'begin transaction' statement.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@575 f10b8b52-1f0f-0410-b9dc-dace7d27bc78