Revert to using the derived pickling instances (rather than Marshal).
The bug mentioned in the comment for r615 was (at least partly) due to
storing things in the `globals' environment of functions.  I've
changed the type of globals to unit so that it can't be used in that
way.  I didn't remove it altogether because it might be useful when we
want to fix the interactive loop.  Perhaps it would be better to
remove it altogether.

The pickled value is still pretty large.  Perhaps we should take
advantage of sharing instead of expanding values in place during
pickling.  It seems worth investigating.  If that's the way to go then
the stuff in monadic_pickle.ml is a reasonable starting point.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@616 f10b8b52-1f0f-0410-b9dc-dace7d27bc78