Fixes for type inference bugs (reported by Thierry and Jeremy).
The bugs were cases of the same thing: the creation of a datatype that
wasn't wrapped in a `MetaTypeVar. Ensuring that `TypeVar,
`RigidTypeVar and, in particular, `Recursive are always wrapped in a
`MetaTypeVar can be quite subtle. We should consider refactoring in
order to enable the type system to give us a bit more help here.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@650 f10b8b52-1f0f-0410-b9dc-dace7d27bc78