Fix the problem wherein the main process/event handlers run out of stack.
The problem arises because event handlers don't yield, and our CPS
code makes lots of tail calls, so we soon run out of stack.  We can't
call setTimeout to clear the stack, because that would allow other
event handlers to run, breaking the model.  Instead, here's a
trampoline, of sorts, which clears the stack in a different way.

The first call to _yield starts up the trampoline: it sets up a
try/catch block then calls the continuation.  A while later, at the
point where we'd call setTimeout if we were in another process we
throw an exception containing the continuation.  The try/catch set up
earlier catches the exception, jumps back to the beginning of the
(try/catch) block and invokes the continuation.  Throwing the
exception threw away the stack, of course, so we can continue in this
vein ad infinitum.

There's still a problem with mandelcolor: now the changes to the DOM
don't actually update the screen until the end of the computation, but
I think that's a separate problem.




git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@702 f10b8b52-1f0f-0410-b9dc-dace7d27bc78