Add a sample testing framework.  For now, just the list tests.
Feedback / comments very welcome!

There's a bug because there's no reasonable way to compare types at
present.  I expect the right way is to have a canonical form for types
(minimised, then with type variables numbered according to some
order); for now just make the (wrong) assumption that equivalence of
types is the same as equivalence of printed representations.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@706 f10b8b52-1f0f-0410-b9dc-dace7d27bc78