Various build changes.
Update to the latest OCamlMakefile.

Change our Makefile around a bit: remove some cruft, and move the
deriving preprocessor information into the source files: you put

  (*pp deriving *)

at the top of the source file to indicate that it should be
preprocessed with "deriving".  This actually makes things a bit
faster, since only the files that need it are preprocessed.

Switch to using the optimised ocamldep.

Change type_basis.ml from DOS to Unix line endings.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@711 f10b8b52-1f0f-0410-b9dc-dace7d27bc78