Remove a limitation that's been bothering me for a while: it wasn't possible to write a function that's polymorphic over typed and untyped expressions and access the position data.  The reason was that the typed and untyped data had no intersection, so you couldn't write a function that was polymorphic over *that*.
One way to fix that would be to use objects to store all data, using
methods that were in the signatures of both typed and untyped data
objects for access.  That's not ideal, though, since objects are quite
cumbersome to use: they don't allow pattern matching, for example.

Instead of objects (which are essentially polymorphic records), we can
use the dual, polymorphic variants, tagging typed and untyped data as
such and having certain functions (e.g. Syntax.position) that accept
both tags.  We can now mark function as accepting

   * only typed expressions
   * only untyped expressions
   * both typed and untyped expressions
   * any kind of expression

which seems pretty optimal.  The syntactic overhead is pretty low, as
well.  Functions that accept both typed and untyped expressions can
access expression data in the intersection (i.e. positions).

There are actual use cases for this: it'd be nice if
Interpreter.interpret accepted either typed or untyped expressions,
for example.  (Not there yet, but that's because other data structures
(continuations) assume typed expressions.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@726 f10b8b52-1f0f-0410-b9dc-dace7d27bc78