- fix bug in setting PATH In a makefile x += v appends " v", i.e. a space as well as v. Use x := $(x)v instead.  - use ocamldep instead of ocamldep.opt by default ocamldep.opt isn't included with all distributions of OCaml (e.g. the MinGW port)
I've now persuaded links to compile under the MinGW version of
OCaml 3.09.3. It isn't feasible to get the the MSVC version
working yet as the OCaml developers don't support the latest
version of Visual Studio. Also, attempting to compile without
Cygwin on Windows is a waste of time as none of the
other Unix toolsets are anywhere near properly supported
(e.g. UnxTools hasn't been touched for the last three years).

At least having a MinGW version means we can create a
standalone Windows binary, although I'm not sure how easy
it is to link this with a non-cygwin database library.




git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@730 f10b8b52-1f0f-0410-b9dc-dace7d27bc78