Implemented type aliases
  typename Foo = t

defines Foo as an alias for t.

  typename Bar(a1, ..., an) = t

defines Bar(t1, ..., tn) as an alias for t[a1 := t1, ..., an := tn],
for all t1, ..., tn.

Issues:

 - Type aliases conflict with our previous implementation of abstract
types. The current implementation does not allow new abstract types to
be defined.

 fun (x:Foo) {x}

defines a function of type Foo -> Foo, where Foo is a type alias - not
an abstract type.

Currently the only abstract type is Event, which is bound to an alias
in the initial alias environment defined in library.ml.

 - We should amend the pretty printer to output Foo instead of Foo () as
it does now for nullary type aliases.

 - The type checker does not check for existence of aliases until they
are unified.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@740 f10b8b52-1f0f-0410-b9dc-dace7d27bc78