Refactored type inference code to use a global mailbox variable (as described in our ESOP 07 submission) instead of threading the current mailbox through all abstractions.
This required propagating rigid type variables throughout the compiler
in order to be able to distinguish rigid and flexible type variables
in type annotations.

e.g. the quantified type of:
  fun (x) {x} : a -> a
is (in the abstract syntax):
  ([`RigidTypeVar "a"; `TypeVar "_1"],
   FunctionType (RigidTypeVar "a", TypeVar "_1", RigidTypeVar "a"))

For consistency, we should probably add rigid row variables throughout
as well.

We should also think about what syntax to use for explicit
quantification over rigid and flexible types.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@750 f10b8b52-1f0f-0410-b9dc-dace7d27bc78