- changed Record_intro to take a map from labels to expressions instead of a list of (label, expression) pairs  - implemented optimisations (e.g. take/drop) for Record_intro
Next step is to get rid of Record_extension by merging it into
Record_intro.

For some reason the take/drop optimisation still isn't happening in
conjunction with joins (e.g. in the winestore). (Note that this has
been broken for a long time - it isn't even happening in the version
we have on the public website.)



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@813 f10b8b52-1f0f-0410-b9dc-dace7d27bc78