More comprehensive analysis for elimination of unused functions before generating JS.  This should remove all functions that are unused in the ("static") callgraph, including recursive (and mutually recursive) functions.
This change also exposes

  1. a bug (that Syntax.freevars doesn't consider l:name bindings)

  2. and a sort-of bug (that "event" is treated specially
     by the type inference module.).

(which means that some of the demos are temporarily broken.)

The first is a bit unpleasant, but I'll check a fix in soon.  What's
the right fix for the second?  Perhaps making "event" globally visible
is the most reasonable thing to do for now.  I don't think treating
"event" specially in Callgraph.elim_dead_defs is the right thing to
do.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@863 f10b8b52-1f0f-0410-b9dc-dace7d27bc78