Remove l:name during desugaring so we don't have to deal with it in the rest of the compiler.
During desugaring:

   * l:name is replaced by name

   * references to l:name-bound variables in the l:action are replaced
     by explicit environment lookups

   * references to l:name-bound variables in l:on-foo handlers are
     replaced by calls to getInputValue

I haven't tested this heavily, but it seems to work.

We should do something similar for the "event" value in handlers.

The one thing I'm not sure about is the removal of the "onfocus" bits
and pieces that js.ml inserts for elements that have l:name.  Do we
still make use of that anywhere?  Is what it did the right thing to
do?

Also fix broken tests from r1103.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@1105 f10b8b52-1f0f-0410-b9dc-dace7d27bc78