Replace List.assoc with a version that raises
    NotFound v

instead of

    Not_found

if the first argument is a string.  (This is awful, and uses the
dreaded Obj module.  Perhaps it should be turned off by default and
only used in emergencies.)

Also fix Set.max_elt, Set.min_elt and Set.choose to raise

    Invalid_argument("Set.max_elt")

etc. instead of

    Not_found

if passed an empty set.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@1311 f10b8b52-1f0f-0410-b9dc-dace7d27bc78