Fixed a bug where aliases weren't being expanded.
This required separating out concrete_types into two functions: one
for just expanding MetaTypeVars and another that in addition expands
aliases. Furthermore, the expand_aliases function had to be moved into
a separate module as it needs to call Instantiate.aliases which
depends on the Types module.

Also got rid of the blinking and random colours in type error
messages.



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@1428 f10b8b52-1f0f-0410-b9dc-dace7d27bc78