- Added the 'identity' constraint for tables (aka auto_increment).  - Implemented basic insert ... returning ... functionality.
Currently the returning clause must be a variable name. The type of
tables should probably be augmented with an optional identity
field. Then we can actually type-check insert ... returning .... I'll
probably leave this to coincide with when I overhaul the table
type to support needed fields and defaults.

The expression

  insert ... returning id

compiles in mysql to

  insert ...;
  select last_insert_id()

and in postgres 8.2 and later to

  insert ... returning id

It is not yet implemented for sqlite (should be easy) or postgres
versions prior to 8.2 (which don't support insert ... returning ...).



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@1453 f10b8b52-1f0f-0410-b9dc-dace7d27bc78