This is disabled by default, so it shouldn't break existing stuff. To enable it switch on the 'quantified_instantiation' setting.
Partial support for 'first-class' polymorphism. There are still some
issues that need addressing before this actually works properly. I
think type inference more-or-less works, but the generated IR is
definitely wrong in some cases. The difficulty is in generating the
correct type abstractions and applications (see the bug description in
typeSugar.ml).



git-svn-id: http://groups.inf.ed.ac.uk/links/svn/trunk@2001 f10b8b52-1f0f-0410-b9dc-dace7d27bc78